Add PHA2 dataset
The files are from the http://tgcat.mit.edu/ processing of obsid 16221
by Chandra. The target is 3C120. The files have been changed from the
TGCAT-provided files by prepending '3c120_' to the file names and,
for the *pha files, the ANCRFILE and RESPFILE keywords have been updated
to match this new scheme.