Use physical not WCS coordinates for lev3fft test
It is not obvious what exactly this test is checking (maybe it
is the WCS support, but I doubt it), but we know that fitting with
WCS is sub-optimal (particularly with convolution, as used here
with PSFs).

I have changed the fits to use the physical rather than WCS
coord system. It means that the FWHM value is somewhat more
meaningful (pixels, not degrees), and [not obvious here, but
will be in the corresponding sherpa update] ends up with a
positive background model.